Added switch statement and skeleton code for Fr94 Indexer
Added a switch statement for us to use where we just have to create a specific indexer object for the document type and then send the document to add fields etc. to it.

TODO: Actually fill in Fr94 documents properly with class functions.

I also added a command parameter in run.sh so we can run it using that.